anomaly-detector/processor/requirements.txt
requests
kafka-python==2.2.10
prometheus_client>=0.18.0
PyYAML==6.0.2
